+++ CHANGELOG.md
@@ -26,9 +26,16 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 -->
 ------
+## [v10.0.2](https://github.com/asfadmin/Discovery-asf_search/compare/v10.0.1...v10.0.2)
+### Fixed
+- `PRODUCTION_CONFIGURATION` constants module now accessible as top level import
+- NISAR Urgent Response works properly with processing level searches
+
+------
 ## [v10.0.1](https://github.com/asfadmin/Discovery-asf_search/compare/v10.0.0...v10.0.1)
 ### Added
 - adds NISAR file sizes to the `nisar` attribute of exports
+
 ------
 ## [v10.0.0](https://github.com/asfadmin/Discovery-asf_search/compare/v9.0.9...v10.0.0)
 ### Added